WIP - Clean and enhance.
