+++ Nysa.CodeAnalysis/Text/Parsing/Building/DownKey.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+using Nysa.Logics;
+using Nysa.Text.Lexing;
+
+namespace Nysa.Text.Parsing.Building;
+
+// Key used to memoize building down.
+internal record struct DownKey(ChartEntry Entry, ChartPosition Position);
