+++ Nysa.CodeAnalysis/Text/Parsing/FinalChart.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
 
 namespace Nysa.Text.Parsing
 {
