+++ Nysa.CodeAnalysis/Text/Parsing/Node.cs
@@ -11,14 +11,13 @@ using Nysa.Text;
 namespace Nysa.Text.Parsing
 {
 
-    public partial class Node : IEquatable<Node>
+    public sealed class Node : IEquatable<Node>
     {
         // instance members
-        public Identifier Id { get; private set; }
-        public String Symbol { get; private set; }
+        public Identifier   Id      { get; init; }
+        public String       Symbol  { get; init; }
 
-        private List<NodeOrToken> _Members;
-        public IReadOnlyList<NodeOrToken> Members { get => this._Members; }
+        public IReadOnlyList<NodeOrToken> Members { get; init; }
 
         public Option<TextSpan> First()
             => this.Members.Count == 0 ? Option<TextSpan>.None : this.Members.Select(m => m.Match(a => a.First(), b => b.Span.Some())).First();
@@ -31,24 +30,24 @@ namespace Nysa.Text.Parsing
         {
             this.Id         = id;
             this.Symbol     = symbol;
-            this._Members   = members.AllNotNull().ToList();
+            this.Members    = members.AllNotNull().ToList();
         }
 
         public Node(Identifier id, String symbol, IEnumerable<NodeOrToken> members)
         {
             this.Id         = id;
             this.Symbol     = symbol;
-            this._Members   = members.ToList();
+            this.Members    = members.ToList();
         }
 
         public Boolean Equals(Node? other)
-            =>    !Object.ReferenceEquals(other, null)
+            =>    other != null
                && this.Id == other.Id
-               && this._Members.Count == other._Members.Count
-               && this._Members.Zip(other.Members, (f, s) => f.Equals(s)).All(t => t);
+               && this.Members.Count == other.Members.Count
+               && this.Members.Zip(other.Members, (f, s) => f.Equals(s)).All(t => t);
 
         public override Int32 GetHashCode()
-            => this.Id.HashWithOther(this._Members.HashAll());
+            => this.Id.HashWithOther(this.Members.HashAll());
 
         public override string ToString() => $"<{this.Symbol}>";
